+++ cmd/security-profiles-operator/main.go
@@ -35,11 +35,11 @@ import (
 	seccompprofilev1alpha1 "sigs.k8s.io/security-profiles-operator/api/seccompprofile/v1alpha1"
 	selinuxpolicyv1alpha1 "sigs.k8s.io/security-profiles-operator/api/selinuxpolicy/v1alpha1"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/config"
-	"sigs.k8s.io/security-profiles-operator/internal/pkg/controllers/profilerecorder"
-	"sigs.k8s.io/security-profiles-operator/internal/pkg/controllers/seccompprofile"
-	"sigs.k8s.io/security-profiles-operator/internal/pkg/controllers/selinuxpolicy"
-	"sigs.k8s.io/security-profiles-operator/internal/pkg/controllers/spod"
-	"sigs.k8s.io/security-profiles-operator/internal/pkg/enricher"
+	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/enricher"
+	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/profilerecorder"
+	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/seccompprofile"
+	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/selinuxpolicy"
+	"sigs.k8s.io/security-profiles-operator/internal/pkg/manager/spod"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/nonrootenabler"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/version"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/webhooks/binding"
