+++ deploy/base/webhook/deployment.yaml
@@ -16,7 +16,7 @@ spec:
       labels:
         name: security-profiles-operator-webhook
     spec:
-      serviceAccountName: security-profiles-operator
+      serviceAccountName: spo-webhook
       containers:
         - name: security-profiles-operator
           image: security-profiles-operator
