+++ deploy/base/webhook/service_account.yaml
@@ -1,0 +1,6 @@
+---
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: spo-webhook
+  namespace: security-profiles-operator
