+++ internal/pkg/daemon/selinuxpolicy/setup.go
