+++ internal/pkg/manager/spod/bindata/default_profiles.go
