+++ internal/pkg/manager/spod/bindata/spod.go
@@ -64,7 +64,7 @@ var Manifest = &appsv1.DaemonSet{
 				},
 			},
 			Spec: v1.PodSpec{
-				ServiceAccountName: config.OperatorName,
+				ServiceAccountName: config.SPOdServiceAccount,
 				InitContainers: []v1.Container{
 					{
 						Name:            "non-root-enabler",
